Add deploy.sh script
Adds a Bash script to deploy website to remote server. The script can be
invoked like so

    $ ./deploy.sh remote_user remote_port

where `remote_user` is the appropriate user account with SSH access on
the remote server and `remote_port is appropriate port for SSH on the
remote server. For security reasons, these actual values are omitted
from this repo.